+++ src/components/Login.css
@@ -1,0 +1,16 @@
+.login-form {
+    display: flex;
+    flex-direction: column;
+    background-color: lightgray;
+    border-radius: 20px;
+    margin-top: 2rem;
+    justify-content: center;
+    align-items: center;
+  }
+  .login-form > * {
+    margin-top: 2rem;
+  }
+  .login-btn {
+    margin-bottom: 2rem;
+    width: 6rem;
+  }
