+++ src/index.js
@@ -2,12 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import { BrowserRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
- <React.StrictMode>
-    <App />
- </React.StrictMode>
+ <BrowserRouter>
+  <App />
+ </BrowserRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
